requirements.txt
# Backend Dependencies - ALL FREE!
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
numpy==1.26.3
pydantic==2.5.3
pytesseract==0.3.10  # FREE OCR - no API keys needed!
Pillow==10.2.0
httpx==0.26.0

# Note: You also need Tesseract OCR installed on your system:
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Linux:   sudo apt install tesseract-ocr
# Mac:     brew install tesseract
